fix(profile stepper): add profile picture placeholder & rename files
