+++ front/src/components/ProfilePicture.tsx
@@ -1,17 +1,24 @@
 import { Avatar } from '@mantine/core';
+import { IconUserCircle } from '@tabler/icons-react';
 
 const ProfilePicture = ({
   src,
   radius = 'xl',
   size = 'xl',
   alt,
+  placeholderIcon = false,
 }: {
   src: string | null | undefined;
   radius?: number | string;
   size?: number | string;
   alt: string | null | undefined;
+  placeholderIcon?: boolean;
 }) => {
-  return <Avatar radius={radius} size={size} src={src} alt={alt ?? ''} />;
+  return (
+    <Avatar radius={radius} size={size} src={src} alt={alt ?? ''}>
+      {placeholderIcon && <IconUserCircle size="3rem" />}
+    </Avatar>
+  );
 };
 
 export default ProfilePicture;
