+++ front/src/configs/steppers/stepperArtistConfig.ts
@@ -1,4 +1,4 @@
-import { LinkUtils } from '../services/LinkUtils';
+import { LinkUtils } from '../../services/LinkUtils';
 export const artistInitialValues = {
   name: '',
   description: '',
